Fix the argument parsing issue when the original script needs complex args
